refactor gnosis code
